pyproject.toml
[project]
name = "vat-filemaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=43.0.0",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyside6>=6.10.0",
    "pyside6-fluent-widgets[full]>=1.9.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
]
